requirements.txt
streamlit==1.16.0
google-generativeai==0.1.0
python-dotenv==0.21.0
docx==0.2.8
